Replace lists:last/1 with erlang:hd/1
Kannel splits long messages and sets registered_delivery flag for the first part only.
For both Kannel and NowSMS common parts are the same, so from current point of view
this replacement is safe. If it will turn out to be wrong we have to turn it back and
implement more meaningfull merge to fetch the registered_delivery flag